docs(exploring_the_field): add dichotomy to refinement step
Image used in the description is a drawio html page allowing some interactions with layers.

Relevant documentation to add interactions with drawio:

- https://www.drawio.com/blog/interactive-diagram-layers
- https://www.drawio.com/doc/faq/custom-links

To include it I used the `raw` directive but I didn’t find a way to add
a caption like it would have been possible with a image so I used an
admonition instead.

Refs: #68