fix(common): save not georeferenced output
In case images are not georeferenced there is no transformation thus we
can not change the ROI origin nor set pixel size in case of use of step
in calculating matching cost.

At first I would simply add an identity transformation ideally in
`create_dataset_from_inputs` from Pandora but it seems that it was done
in previous versions of Pandora and caused troubles with QGIS (see
3d/PandoraBox/pandora#211). I also found a [bug report to
QGIS](https://github.com/qgis/QGIS/issues/45324) but it is closed for
inactivity.

Refs: #164, 3d/PandoraBox/pandora#211