chore(install): use meson-python to build cpp
Following this:
https://mesonbuild.com/meson-python/how-to-guides/editable-installs.html

In order to use the editable mode we need to add `meson-python meson
ninja` dependencies to our virtual environment and use
`--no-build-isolation` option on installation.

Following this:
https://mesonbuild.com/Using-wraptool.html#automatic-dependency-fallback

We can download wraps to download dependencies with this command:

```commandline
meson wrap update-db
```

It download in the `subprojects` directory thus we add it to `.gitignore`.

To run cpp tests:

```commandline
meson test -C build/cp311/
```

Other references:

https://mesonbuild.com/Python-module.html

https://mesonbuild.com/Dependencies.html

https://mesonbuild.com/Syntax.html