fix(disp_map): make datavars data independent
Before: all datavars shared the same data array. Thus, a change in one
datavar was also reflected in the others.

Now: each datavar has its own data. Note that memory consumption is
increased (this is how the bug was detected: the consumption was too low
compared to the expected usage).