refactor(state_machine): extract shared transition logic into abstract base class
Transitions are expected to be identical between CheckMachine and Pandora2DMachine.
This commit enforces that constraint by introducing a common abstract class.

The only divergence lies in the need for callbacks during the
"matching_cost" step in Pandora2DMachine. This is handled by invoking
`add_callback` method of the corresponding transitions within the
`__init__` method of Pandora2DMachine.