feat(common): save dataset with nodata tif attribute and band name set
Save the nodata attribute of disparity maps with `invalid_disp` value.

The function that was used to save tif was a function from Pandora. As
we do not want to modify Pandora project, we reimplement it in
Pandora2d.

We set band names with dataarray names when there is only one band.

feat: add nodata to correlation_score.tif

correlation_score is initialized with the `invalid_disp` value, which
still appears along the edges of the map in the TIFF file. Therefore,
this value should be treated as `nodata` in the TIFF metadata.